feat: Phase-9 design alignment — Debezium CDC + Avro/CloudEvents + inbox + minor units + SASL
Replace JSON/app-level-outbox with the documented Phase-9 architecture:

P-1: Docker build fixes (repo-root context, PostgreSQL WAL=logical, CDC role,
     explicit Kafka topics); all 5 images build clean.
P0:  20 CloudEvents-wrapped Avro schemas for full catalog (BACKWARD default);
     Schema Registry integration + CI validation job + registration scripts.
P1:  Producer rearchitecture — CDC outbox table (design shape) per service;
     Debezium connectors w/ EventRouter SMT + AvroConverter → new domain topics;
     live spike verified outbox→Debezium→Avro on payments.payment.created.
P2:  Consumer rearchitecture — Avro deserialize (Confluent/registry) in 3
     languages; inbox pattern (consumer_inbox) + retry scheduler (exp backoff
     ≤5) + per-domain DLQ (payments.dlq/ledger.dlq/notifications.dlq).
P3:  All amounts → long minor units (cents): DTOs, entities, ledger math, DB
     columns; ledger sum-zero verified with integer arithmetic.
P4:  Kafka SASL_PLAINTEXT/PLAIN + Schema Registry HTTP BASIC auth; env-driven
     with PLAINTEXT default (dev-only credentials).
P5:  Decommissioned old JSON topics, processed_events, *_outbox tables, legacy
     pollers, obsolete contracts; updated README/verify scripts/CI (Go 1.25).

Preserved the serial payment→fraud→ledger→notification chain (OQ-C).

+++ services/java/payment-service/src/main/java/com/paymentapi/paymentservice/dto/PaymentResponse.java
@@ -1,14 +1,13 @@
 package com.paymentapi.paymentservice.dto;
 
 import com.paymentapi.paymentservice.entity.Payment;
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
 public record PaymentResponse(
     UUID paymentId,
     String status,
-    BigDecimal amount,
+    long amount,   // minor currency units (cents)
     String currency,
     String merchantId,
     String customerId,
